migrate examples to sandpack
