+++ docs/examples/table/topCaptionExample.js
@@ -1,0 +1,106 @@
+// @flow strict
+import { type Node } from 'react';
+import { Table, Box, Label, Checkbox, Text, Status, Flex } from 'gestalt';
+
+function HeaderRow({ id }: any) {
+  return (
+    <Table.Header>
+      <Table.Row>
+        <Table.HeaderCell>
+          <Box display="visuallyHidden">
+            <Label htmlFor={id}>Not all checkboxes are checked</Label>
+          </Box>
+          <Checkbox id={id} onChange={() => {}} indeterminate size="sm" />
+        </Table.HeaderCell>
+        {['Status', 'Campaign'].map((title) => (
+          <Table.HeaderCell key={title}>
+            <Text weight="bold">{title}</Text>
+          </Table.HeaderCell>
+        ))}
+      </Table.Row>
+    </Table.Header>
+  );
+}
+
+function BaseRow({ id, checked, disabled, type, text, subtext, campaign }: any) {
+  return (
+    <Table.Row>
+      <Table.Cell>
+        <Checkbox
+          id={`${id.replace(/ /g, '_').replace(/'/g, '')}_${text
+            .replace(/ /g, '_')
+            .replace(/'/g, '')}`}
+          onChange={() => {}}
+          disabled={disabled}
+          size="sm"
+          checked={checked}
+        />
+      </Table.Cell>
+      <Table.Cell>
+        <Status type={type} title={text} subtext={subtext} />
+      </Table.Cell>
+      <Table.Cell>
+        <Label
+          htmlFor={`${id.replace(/ /g, '_').replace(/'/g, '')}_${text
+            .replace(/ /g, '_')
+            .replace(/'/g, '')}`}
+        >
+          <Text color={disabled ? 'gray' : 'darkGray'}>{campaign}</Text>
+        </Label>
+      </Table.Cell>
+    </Table.Row>
+  );
+}
+
+export default function Example(): Node {
+  const tableID = 'Example of correct accessibility with top caption';
+
+  return (
+    <Box padding={4}>
+      <Flex gap={{ column: 2, row: 0 }} direction="column">
+        <Box aria-hidden>
+          <Text size="400" weight="bold">
+            Your Campaigns Summary
+          </Text>
+        </Box>
+        <Table accessibilityLabel="Your campaigns summary">
+          <HeaderRow id={tableID} />
+          <Table.Body>
+            <BaseRow
+              id={tableID}
+              checked
+              type="inProgress"
+              text="In progress"
+              subtext="Ends 11/20/2021"
+              campaign="Engagement"
+            />
+            <BaseRow
+              id={tableID}
+              disabled
+              type="halted"
+              text="Paused"
+              subtext="Ends 11/20/2021"
+              campaign="Awareness"
+            />
+            <BaseRow
+              id={tableID}
+              checked
+              type="warning"
+              text="Warning"
+              subtext="Ends 11/20/2021"
+              campaign="Catalogs"
+            />
+            <BaseRow
+              id={tableID}
+              checked={false}
+              type="ok"
+              text="Complete"
+              subtext="Ends 11/20/2021"
+              campaign="Awareness"
+            />
+          </Table.Body>
+        </Table>
+      </Flex>
+    </Box>
+  );
+}
